add custom asset-listing resolver class
