+++ src/GraphQL/Resolver/Element.php
@@ -17,25 +17,14 @@ namespace Pimcore\Bundle\DataHubBundle\GraphQL\Resolver;
 
 use GraphQL\Type\Definition\ResolveInfo;
 use GraphQL\Type\Definition\UnionType;
-use Pimcore\Bundle\DataHubBundle\Configuration;
-use Pimcore\Bundle\DataHubBundle\Event\GraphQL\Model\ListingEvent;
-use Pimcore\Bundle\DataHubBundle\GraphQL\ElementDescriptor;
 use Pimcore\Bundle\DataHubBundle\GraphQL\Exception\ClientSafeException;
-use Pimcore\Bundle\DataHubBundle\GraphQL\Exception\NotAllowedException;
-use Pimcore\Bundle\DataHubBundle\GraphQL\Helper;
 use Pimcore\Bundle\DataHubBundle\GraphQL\Traits\PermissionInfoTrait;
 use Pimcore\Bundle\DataHubBundle\GraphQL\Traits\ServiceTrait;
 use Pimcore\Bundle\DataHubBundle\GraphQL\Service;
 use Pimcore\Bundle\DataHubBundle\GraphQL\FieldHelper\AbstractFieldHelper;
-use Pimcore\Bundle\DataHubBundle\WorkspaceHelper;
-use Pimcore\Db;
-use Pimcore\Model\Asset;
 use Pimcore\Model\DataObject\AbstractObject;
-use Pimcore\Model\Document;
-use Pimcore\Model\Listing\AbstractListing;
 use Pimcore\Model\Property;
 use Pimcore\Model\Element\Service as ElementService;
-use Pimcore\Bundle\DataHubBundle\PimcoreDataHubBundle;
 
 
 class Element
@@ -155,156 +144,9 @@ class Element
      * @param ResolveInfo|null $resolveInfo
      * @return mixed
      */
-    public function resolveEdge($value = null, $args = [], $context, ResolveInfo $resolveInfo = null)
-    {
-        $element = $value['node'];
-
-        if (null !== $element) {
-            return $this->extractSingleElement($element, $args, $context, $resolveInfo);
-        }
-
-        return null;
-    }
-
-    /**
-     * @param null $value
-     * @param array $args
-     * @param array $context
-     * @param ResolveInfo|null $resolveInfo
-     * @return mixed
-     */
     public function resolveEdges($value = null, $args = [], $context, ResolveInfo $resolveInfo = null)
     {
         return $value['edges'];
-    }
-
-    /**
-     * @param null $value
-     * @param array $args
-     * @param array $context
-     * @param ResolveInfo|null $resolveInfo
-     * @return array
-     * @throws \Exception
-     */
-    public function resolveListing($value = null, $args = [], $context, ResolveInfo $resolveInfo = null)
-    {
-        if ($args && isset($args['defaultLanguage'])) {
-            $this->getGraphQlService()->getLocaleService()->setLocale($args['defaultLanguage']);
-        }
-
-        $modelFactory = $this->getGraphQlService()->getModelFactory();
-        $listClass = $this->getListType();
-
-        /** @var AbstractListing $objectList */
-        $objectList = $modelFactory->build($listClass);
-        $conditionParts = [];
-        if (isset($args['ids'])) {
-            $conditionParts[] = '(o_id IN (' . $args['ids'] . '))';
-        }
-
-        // paging
-        if (isset($args['first'])) {
-            $objectList->setLimit($args['first']);
-        }
-
-        if (isset($args['after'])) {
-            $objectList->setOffset($args['after']);
-        }
-
-        // sorting
-        if (!empty($args['sortBy'])) {
-            $objectList->setOrderKey($args['sortBy']);
-            if (!empty($args['sortOrder'])) {
-                $objectList->setOrder($args['sortOrder']);
-            }
-        }
-
-        // Include unpublished
-//        if (isset($args['published']) && $args['published'] === false) {
-//            $objectList->setUnpublished(true);
-//        }
-
-        /** @var $configuration Configuration */
-        $configuration = $context['configuration'];
-//        $sqlListCondition = $configuration->getSql();
-//
-//        if ($sqlListCondition) {
-//            $conditionParts[] = '(' . $sqlListCondition . ')';
-//        }
-
-        // check permissions
-        $db = Db::get();
-        $tableName = $this->getTypeTableName();
-        $workspacesTableName = 'plugin_datahub_workspaces_' . $this->elementType;
-        $conditionParts[] = ' (
-            (
-                SELECT `read` from ' . $db->quoteIdentifier($workspacesTableName) . '
-                WHERE ' . $db->quoteIdentifier($workspacesTableName) . '.configuration = ' . $db->quote($configuration->getName()) . '
-                AND LOCATE(CONCAT(' . $db->quoteIdentifier($tableName) . '.path,' . $db->quoteIdentifier($tableName) . '.filename),' . $db->quoteIdentifier($workspacesTableName) . '.cpath)=1
-                ORDER BY LENGTH(' . $db->quoteIdentifier($workspacesTableName) . '.cpath) DESC
-                LIMIT 1
-            )=1
-            OR
-            (
-                SELECT `read` from ' . $db->quoteIdentifier($workspacesTableName) . '
-                WHERE ' . $db->quoteIdentifier($workspacesTableName) . '.configuration = ' . $db->quote($configuration->getName()) . '
-                AND LOCATE(' . $db->quoteIdentifier($workspacesTableName) . '.cpath,CONCAT(' . $db->quoteIdentifier($tableName) . '.filename,' . $db->quoteIdentifier($tableName) . '.filename))=1
-                ORDER BY LENGTH(' . $db->quoteIdentifier($workspacesTableName) . '.cpath) DESC
-                LIMIT 1
-            )=1
-        )';
-
-        if (isset($args['filter'])) {
-            $filter = json_decode($args['filter'], false);
-            if (!$filter) {
-                throw new ClientSafeException('unable to decode filter');
-            }
-            $filterCondition = Helper::buildSqlCondition($tableName, $filter);
-            $conditionParts[] = $filterCondition;
-        }
-
-        if ($conditionParts) {
-            $condition = implode(' AND ', $conditionParts);
-            $objectList->setCondition($condition);
-        }
-
-//        $event =  new ListingEvent(
-//            $objectList,
-//            $args,
-//            $context,
-//            $resolveInfo
-//        );
-//        $this->eventDispatcher->dispatch(ListingEvents::PRE_LOAD, $event);
-//        $objectList = $event->getListing();
-
-        $totalCount = $objectList->getTotalCount();
-        $objectList = $objectList->load();
-
-        $nodes = [];
-
-        foreach ($objectList as $element) {
-            $nodes[] = [
-                'cursor' => $this->elementType . '-' . $element->getId(),
-                'node' => $element,
-            ];
-        }
-        $connection = [];
-        $connection['edges'] = $nodes;
-        $connection['totalCount'] = $totalCount;
-
-        return $connection;
-    }
-
-    /**
-     * @param null $value
-     * @param array $args
-     * @param array $context
-     * @param ResolveInfo|null $resolveInfo
-     * @return mixed
-     */
-    public function resolveListingTotalCount($value = null, $args = [], $context, ResolveInfo $resolveInfo = null)
-    {
-        return $value['totalCount'];
     }
 
     /**
@@ -318,57 +160,6 @@ class Element
             $arguments[] = isset($args['objectTypes']) ? $args['objectTypes'] : [AbstractObject::OBJECT_TYPE_OBJECT, AbstractObject::OBJECT_TYPE_FOLDER];
         }
         return $arguments;
-    }
-
-    /**
-     * @param array $elements
-     * @param array $args
-     * @param $context
-     * @param ResolveInfo|null $resolveInfo
-     * @return array
-     * @throws \Exception
-     */
-    protected function extractMultipleElements($elements, $args, $context, $resolveInfo)
-    {
-        $result = [];
-        if ($elements) {
-            foreach ($elements as $element) {
-                $result[] = $this->extractSingleElement($element, $args, $context, $resolveInfo);
-            }
-        }
-        return array_filter($result);
-    }
-
-    /**
-     * @param Element $element
-     * @param array $args
-     * @param $context
-     * @param ResolveInfo|null $resolveInfo
-     * @return array
-     * @throws \Exception
-     */
-    protected function extractSingleElement($element, $args, $context, $resolveInfo)
-    {
-        // Check Workspace permissions
-        if (!WorkspaceHelper::isAllowed($element, $context['configuration'], 'read')) {
-            if (PimcoreDataHubBundle::getNotAllowedPolicy() == PimcoreDataHubBundle::NOT_ALLOWED_POLICY_EXCEPTION) {
-                throw new NotAllowedException('not allowed to view ' . $element->getFullPath());
-            } else {
-                return null;
-            }
-        }
-
-        $data = new ElementDescriptor($element);
-        $data['id'] = $element->getId();
-
-        // Check element type
-        $treeType = $this->getTreeType();
-        $elementType = $treeType->resolveType($data, $context, $resolveInfo);
-        if (in_array($elementType, $treeType->getTypes(), true)) {
-            $this->getFieldHelper()->extractData($data, $element, $args, $context, $resolveInfo);
-            return $data;
-        }
-        return null;
     }
 
     /**
@@ -386,40 +177,6 @@ class Element
             default:
                 trigger_error("unknown element type");
         }
-        return null;
-    }
-
-    /**
-     * @return ?string
-     */
-    protected function getListType()
-    {
-        switch ($this->elementType) {
-            case 'asset':
-                return Asset\Listing::class;
-            case 'document':
-                return Document\Listing::class;
-            default:
-                trigger_error("unknown element type");
-        }
-
-        return null;
-    }
-
-    /**
-     * @return ?string
-     */
-    protected function getTypeTableName()
-    {
-        switch ($this->elementType) {
-            case 'asset':
-                return 'assets';
-            case 'document':
-                return 'documents';
-            default:
-                trigger_error("unknown element type");
-        }
-
         return null;
     }
